[TASK-3] distributed everything into packages
