+++ aleksey.maslov/task-3/cmd/service/main.go
@@ -1,111 +1,31 @@
 package main
 
 import (
-	"encoding/xml"
 	"flag"
-	"fmt"
-	"os"
-	"sort"
-	"strconv"
-	"strings"
 
-	"gopkg.in/yaml.v3"
+	"github.com/A1exMas1ov/task-3/internal/config"
+	"github.com/A1exMas1ov/task-3/internal/currency"
+	"github.com/A1exMas1ov/task-3/internal/jsonwriter"
+	"github.com/A1exMas1ov/task-3/internal/xmlparser"
 )
-
-type Config struct {
-	InputFile  string `yaml:"input-file"`
-	OutputFile string `yaml:"output-file"`
-}
-
-type CurrencyValue float64
-
-type Valute struct {
-	NumCode  int           `json:"num_code" xml:"NumCode"`
-	CharCode string        `json:"char_code" xml:"CharCode"`
-	Value    CurrencyValue `json:"value" xml:"Value"`
-}
-type ValCurs struct {
-	XMLName xml.Name `xml:"ValCurs"`
-	Valutes []Valute `xml:"Valute"`
-}
-
-func (cv *CurrencyValue) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
-	var str string
-
-	err := d.DecodeElement(&str, &start)
-	if err != nil {
-		return fmt.Errorf("failed to decode currency value: %w", err)
-	}
-
-	str = strings.ReplaceAll(str, ",", ".")
-
-	value, err := strconv.ParseFloat(str, 64)
-	if err != nil {
-		return fmt.Errorf("failed to parse currency value: %w", err)
-	}
-
-	*cv = CurrencyValue(value)
-	return nil
-}
-
-func readConfig(configPath string) (*Config, error) {
-	data, err := os.ReadFile(configPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read config file: %w", err)
-	}
-
-	var config Config
-
-	err = yaml.Unmarshal(data, &config)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse YAML config: %w", err)
-	}
-
-	return &config, nil
-}
-
-func parseXML(path string) ([]Valute, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read XML file: %w", err)
-	}
-
-	var valCurs ValCurs
-	err = xml.Unmarshal(data, &valCurs)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse XML: %w", err)
-	}
-
-	return valCurs.Valutes, nil
-}
-
-func sortByValue(valutes []Valute) {
-	sort.Slice(valutes, func(i, j int) bool {
-		return valutes[i].Value > valutes[j].Value
-	})
-}
-
-func saveJSON(path string, valutes []Valute) error {
-	return nil
-}
 
 func main() {
 	configPath := flag.String("config", "configs/config.yaml", "path to config file")
 	flag.Parse()
 
-	config, err := readConfig(*configPath)
+	config, err := config.ReadConfig(*configPath)
 	if err != nil {
 		panic(err)
 	}
 
-	valutes, err := parseXML(config.InputFile)
+	valutes, err := xmlparser.ParseXML(config.InputFile)
 	if err != nil {
 		panic(err)
 	}
 
-	sortByValue(valutes)
+	currency.SortByValue(valutes)
 
-	err = saveJSON(config.OutputFile, valutes)
+	err = jsonwriter.SaveJSON(config.OutputFile, valutes)
 	if err != nil {
 		panic(err)
 	}
