+++ aleksey.maslov/task-3/go.mod
@@ -2,4 +2,4 @@ module github.com/A1exMas1ov/task-3
 
 go 1.22.7
 
-require gopkg.in/yaml.v3 v3.0.1 // indirect
+require gopkg.in/yaml.v3 v3.0.1
