+++ aleksey.maslov/task-3/internal/jsonwriter/jsonwriter.go
@@ -1,0 +1,9 @@
+package jsonwriter
+
+import (
+	"github.com/A1exMas1ov/task-3/internal/currency"
+)
+
+func SaveJSON(path string, valutes []currency.Valute) error {
+	return nil
+}
